cleanup, contract creation works
